+++ CHANGELOG.md
@@ -5,6 +5,8 @@
   - Improved styling on projects and users pages
   - Improved menu and header display on index pages
   - Improved display of pagination on mobile devices
+- **Adverse Event Changes**
+  - Adverse events are now numbered on the project-level
 - **Project Changes**
   - Improved new project setup workflow
   - Removed `subject_code_name` from project settings to simplify setup